modularizing resolve and string map
